+++ src/main/java/subway/section/domain/AlignSections.java
@@ -1,4 +1,4 @@
-package subway.section;
+package subway.section.domain;
 
 import subway.exceptions.InvalidValueException;
