+++ src/main/java/subway/section/dto/SectionResponse.java
@@ -1,4 +1,4 @@
-package subway.section;
+package subway.section.dto;
 
 public class SectionResponse {
     private Long upStationId;
