+++ src/test/java/subway/line/LineAcceptanceTest.java
@@ -9,8 +9,9 @@ import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import subway.AcceptanceTest;
-import subway.station.Station;
-import subway.station.StationResponse;
+import subway.line.dto.LineRequest;
+import subway.line.dto.LineResponse;
+import subway.station.dto.StationResponse;
 
 import java.util.Arrays;
 import java.util.List;
